1.16.15

+++ Install/CopyFiles.ps1
@@ -1,16 +1,33 @@
 #Set-ExecutionPolicy Bypass -Scope CurrentUser
 
 $cur="$PSScriptRoot"
-$source="$PSScriptRoot\..\MarkdownMonster"
+
+$netfull = $True;    # netfull is full framework - otherwise net core
+if($netfull) {
+    $source="$PSScriptRoot\..\MarkdownMonster\bin\Release\net462"
+}
+else {
+    $source="$PSScriptRoot\..\MarkdownMonster\bin\Release\netcoreapp3.0"
+}    
 $target="$PSScriptRoot\Distribution"
+edpnotify.exe
 
 # delete the Distribution folder
 remove-item -recurse -force ${target}
 
 # copy but exclude libGit extra folders
-robocopy ${source}\bin\Release\net462 ${target} /MIR /XD lib /XF git2*.pdb
+robocopy ${source} ${target} /MIR /XD lib /XD runtimes
 
-robocopy ${source}\bin\Release\net462\lib\win32 ${target}\lib\win32 /MIR /XF git2*.pdb
+if ($netfull) {
+    robocopy ${source}\lib\win32 ${target}\lib\win32 /MIR /XF *.pdb
+    robocopy ${source}\lib\win64 ${target}\lib\win64 /MIR /XF *.pdb
+}
+else {
+    robocopy ${source}\runtimes\win-x86 ${target}\runtimes\win-x86 /MIR 
+    robocopy ${source}\runtimes\win-x64 ${target}\runtimes\win-x64 /MIR 
+    robocopy ${source}\runtimes\win ${target}\runtimes\win /MIR 
+    robocopy ${source}\runtimes\win7 ${target}\runtimes\win7 /MIR
+}
 
 Copy-Item ${cur}\mm.exe ${target}\mm.exe
 Copy-Item ${cur}\license.md ${target}\license.md
@@ -20,11 +37,15 @@ Remove-Item ${target}\*.vshost.*
 Remove-Item ${target}\*.xml
 Remove-Item ${target}\*.user
 Remove-Item ${target}\*.dll.config
-Remove-Item ${target}\.vs -Recurse -Force
+
+if ([System.IO.Directory]::Exists($target + "\.vs")) {
+    Remove-Item ${target}\.vs -Recurse -Force
+}
 
 
 # Roslyn - remove extra files
-Remove-Item ${target}\Addins\Snippets\roslyn -Recurse -Force
+# Remove-Item ${target}\Addins\Snippets\roslyn -Recurse -Force
+
 Remove-Item ${target}\roslyn\Microsoft.CodeAnalysis.VisualBasic.dll
 Remove-Item ${target}\roslyn\Microsoft.DiaSymReader.Native.amd64.dll
 Remove-Item ${target}\roslyn\Microsoft.DiaSymReader.Native.x86.dll
